Retorna media

+++ AvaliadorDisciplinas.API/src/main/java/com/engSoft/entities/Feedback.java
@@ -15,11 +15,11 @@ public class Feedback {
     private Long idCourse;
     private Long idStudent;
     private Long idSemester;
-    private Integer workload;
-    private Integer didactic;
-    private Integer organization;
-    private Integer evaluationSystem;
-    private Integer courseware;
+    private double workload;
+    private double didactic;
+    private double organization;
+    private double evaluationSystem;
+    private double courseware;
     private Util.ModalityEnum modality;
 
     public Feedback(FeedbackDTO feedbackDTO) {
@@ -66,43 +66,43 @@ public class Feedback {
         this.idStudent = idStudent;
     }
 
-    public Integer getWorkload() {
+    public double getWorkload() {
         return workload;
     }
 
-    public void setWorkload(Integer workload) {
+    public void setWorkload(double workload) {
         this.workload = workload;
     }
 
-    public Integer getDidactic() {
+    public double getDidactic() {
         return didactic;
     }
 
-    public void setDidactic(Integer didactic) {
+    public void setDidactic(double didactic) {
         this.didactic = didactic;
     }
 
-    public Integer getOrganization() {
+    public double getOrganization() {
         return organization;
     }
 
-    public void setOrganization(Integer organization) {
+    public void setOrganization(double organization) {
         this.organization = organization;
     }
 
-    public Integer getEvaluationSystem() {
+    public double getEvaluationSystem() {
         return evaluationSystem;
     }
 
-    public void setEvaluationSystem(Integer evaluationSystem) {
+    public void setEvaluationSystem(double evaluationSystem) {
         this.evaluationSystem = evaluationSystem;
     }
 
-    public Integer getCourseware() {
+    public double getCourseware() {
         return courseware;
     }
 
-    public void setCourseware(Integer courseware) {
+    public void setCourseware(double courseware) {
         this.courseware = courseware;
     }
